--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -6,6 +6,5 @@ require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/mailru/easyjson v0.7.7
 	github.com/stretchr/testify v1.7.0
-	golang.org/x/crypto v0.1.0
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )